Fill-a-pix-desktop/requirements.txt
colorama==0.4.4
numpy==1.21.0
Pillow==8.3.1
